requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
exceptiongroup==1.3.1
idna==3.11
iniconfig==2.1.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.2.1
requests==2.32.5
ruff==0.15.6
tomli==2.4.0
typing_extensions==4.15.0
urllib3==2.6.3
